lemonade/ai_social_analysis/requirements.txt
Django==1.9.4
elasticsearch==2.1.0
mysqlclient
numpy
scipy
django-cors-headers

